Move string-utilities to own module
