+++ docs/api/variables.rst
@@ -1,12 +1,11 @@
-.. currentmodule:: pyam.utils
+.. currentmodule:: pyam.str
 
 Variables utilities
 ===================
 
-The **variable** dimension of the |pyam| data format implements
-implements a "semi-hierarchical" structure using the :code:`|` character
-(*pipe*, not l or i) to indicate the *depth*.
-Read the `data model documentation`_ for more information.
+The **variable** dimension of the |pyam| data format implements implements a
+"semi-hierarchical" structure using the :code:`|` character (*pipe*, not l or i)
+to indicate the *depth*. Read the `data model documentation`_ for more information.
 
 .. _`data model documentation`: ../data.html#the-variable-column
 
@@ -16,6 +15,6 @@ The package provides several functions to work with such strings.
 
 .. autofunction:: find_depth
 
-.. autofunction:: reduce_hierarchy
-
 .. autofunction:: get_variable_components
+
+.. autofunction:: reduce_hierarchy
